requirements.txt
argparse==1.2.1
cocos2d==0.6.0
ipython==2.3.1
pyglet==1.2alpha1
PyYAML>=3.11
six==1.8.0
wsgiref==0.1.2
